refactor existing functional test for the submitted callback event
I move test to an specific event folder
rename class
improve test by checking the state is the expected one
